refactor: 참가자 이름 원시값 래핑
