+++ src/main/java/domain/GameResult.java
@@ -1,6 +1,9 @@
 package domain;
 
-import java.util.HashMap;
+import domain.participant.Dealer;
+import domain.participant.Player;
+import domain.participant.Players;
+
 import java.util.LinkedHashMap;
 import java.util.Map;
