+++ src/main/java/domain/Name.java
@@ -1,0 +1,20 @@
+package domain;
+
+public class Name {
+    private String name;
+
+    public Name(String name) {
+        validateNameLength(name);
+        this.name = name;
+    }
+
+    private void validateNameLength(String name) {
+        if (name.length() > 5) {
+            throw new IllegalArgumentException();
+        }
+    }
+
+    public String getName() {
+        return name;
+    }
+}
