+++ src/main/java/view/OutputView.java
@@ -1,7 +1,10 @@
 package view;
 
-import domain.*;
 import domain.card.Card;
+import domain.participant.Dealer;
+import domain.participant.Participant;
+import domain.participant.Player;
+import domain.participant.Players;
 
 import java.util.ArrayList;
 import java.util.List;
