+++ src/test/java/domain/GameResultTest.java
@@ -1,5 +1,7 @@
 package domain;
 
+import domain.participant.Dealer;
+import domain.participant.Players;
 import org.junit.jupiter.api.Test;
 
 import java.util.List;
