+++ src/test/java/domain/NameTest.java
@@ -1,0 +1,15 @@
+package domain;
+
+import org.junit.jupiter.api.Test;
+
+import static org.assertj.core.api.Assertions.assertThatThrownBy;
+import static org.junit.jupiter.api.Assertions.*;
+
+class NameTest {
+    @Test
+    void 이름이_5자를_초과하면_예외를_발생한다() {
+        assertThatThrownBy(() -> new Name("pobiss"))
+                .isInstanceOf(IllegalArgumentException.class);
+    }
+
+}
