+++ src/test/java/domain/ParticipantTest.java
@@ -1,6 +1,8 @@
 package domain;
 
 import domain.card.Card;
+import domain.participant.Participant;
+import domain.participant.Player;
 import org.junit.jupiter.api.BeforeEach;
 import org.junit.jupiter.api.Test;
 import java.util.List;
