+++ src/test/java/domain/ParticipantsTest.java
@@ -1,9 +1,8 @@
 package domain;
 
-import org.junit.jupiter.api.BeforeEach;
+import domain.participant.Players;
 import org.junit.jupiter.api.Test;
 
-import java.util.ArrayList;
 import java.util.List;
 
 import static org.assertj.core.api.Assertions.assertThatThrownBy;
